+++ src/router.tsx
@@ -15,10 +15,10 @@ import HomePage from '@/pages/home/HomePage';
 import WeeklyStatisticsPage from '@/pages/statistics/WeeklyStatisticsPage';
 import MonthlyStatisticsPage from '@/pages/statistics/MonthlyStatisticsPage';
 import PresetSettingPage from '@/pages/PresetSettingPage';
-import MyPage from '@/pages/MyPage';
-import MyPageEditPage from '@/pages/MyPageEditPage';
-import AccountManagePage from '@/pages/AccountManagePage';
-import LeavePage from '@/pages/LeavePage';
+import MyPage from '@/pages/my/MyPage';
+import MyPageEditPage from '@/pages/my/MyPageEditPage';
+import AccountManagePage from '@/pages/my/AccountManagePage';
+import LeavePage from '@/pages/my/LeavePage';
 import HouseWorkStepOnePage from '@/pages/housework/HouseWorkStepOnePage';
 import HouseWorkStepTwoPage from '@/pages/housework/HouseWorkStepTwoPage';
 import GroupInviteReceivePage from '@/pages/group/GroupInviteReceivePage';
